backend/app/config.py
from dotenv import load_dotenv
import os
load_dotenv()
DATABASE_URL = os.getenv("DATABASE_URL", "mysql+pymysql://root:@localhost:3306/FitConnet")
SECRET_KEY = os.getenv("SECRET_KEY", "CAMBIA_POR_ALGO_SEGURO")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60*24
